ces/backend/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.14",
    "deprecated>=1.2.18",
    "google-adk==1.7.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-secret-manager>=2.24.0",
    "google-cloud-texttospeech>=2.25.0",
    "google-genai>=1.20.0",
]
